Updated page_generator file structure
