+++ page_generator.rb
@@ -1,26 +1,26 @@
-class PageGenerator
+require './lib/game'
+require './lib/game_stats'
+require './lib/team'
+require './lib/season'
+require './lib/team_stats'
+require './lib/stat_sorter'
+require './lib/league_stats'
+require './lib/file_converter'
+require './lib/stat_tracker'
+require './lib/page_generator'
+require 'csv'
+require 'erb'
 
-  attr_reader :template,
-              :stat_tracker,
-              :file_path
+games_path = './data/game.csv'
+teams_path = './data/team_info.csv'
+game_teams_path = './data/game_teams_stats.csv'
 
-  def initialize(template, files)
-    @template = File.open(template, 'rb', &:read)
-    @stat_tracker = StatTracker.from_csv(files)
+files = {
+games: games_path,
+teams: teams_path,
+game_teams: game_teams_path
+}
 
-    ENV['SITE'] = ENV['PWD'] + '/site'
-    @file_path = File.join(ENV['SITE'], 'index.html'
-    )
-  end
-
-  def render
-    ERB.new(template).result(binding)
-  end
-
-  def save
-    File.open(file_path, "w+") do |file|
-      file.write(render)
-    end
-  end
-
-end
+template = './site/template.erb'
+page_generator = PageGenerator.new(template, files)
+page_generator.save
